Add note on using a parameter equal to zero as vector multipler
It's feasible to use a parameter equal to zero to specify a bit vector
using a vector multipler plus a vector. Thus it means that it's possible
to define a variable value on a switch() construction based, for
instance, on a variable number of bits proportional to a bus size.